refactor: change checkbox component's path
