+++ src/app/(user)/register/_components/checkbox/index.tsx
